+++ enterprise/coderd/provisionerdaemons_test.go
@@ -13,7 +13,6 @@ import (
 	"cdr.dev/slog/sloggers/slogtest"
 	"github.com/coder/coder/v2/coderd/coderdtest"
 	"github.com/coder/coder/v2/coderd/database"
-	"github.com/coder/coder/v2/coderd/provisionerdserver"
 	"github.com/coder/coder/v2/coderd/rbac"
 	"github.com/coder/coder/v2/coderd/util/ptr"
 	"github.com/coder/coder/v2/codersdk"
@@ -91,7 +90,7 @@ func TestProvisionerDaemonServe(t *testing.T) {
 				codersdk.ProvisionerTypeEcho,
 			},
 			Tags: map[string]string{
-				provisionerdserver.TagScope: provisionerdserver.ScopeOrganization,
+				provisionersdk.TagScope: provisionersdk.ScopeOrganization,
 			},
 		})
 		require.Error(t, err)
@@ -118,7 +117,7 @@ func TestProvisionerDaemonServe(t *testing.T) {
 				codersdk.ProvisionerTypeEcho,
 			},
 			Tags: map[string]string{
-				provisionerdserver.TagScope: provisionerdserver.ScopeOrganization,
+				provisionersdk.TagScope: provisionersdk.ScopeOrganization,
 			},
 		})
 		require.Error(t, err)
@@ -135,7 +134,7 @@ func TestProvisionerDaemonServe(t *testing.T) {
 			},
 		}})
 		closer := coderdtest.NewExternalProvisionerDaemon(t, client, user.OrganizationID, map[string]string{
-			provisionerdserver.TagScope: provisionerdserver.ScopeUser,
+			provisionersdk.TagScope: provisionersdk.ScopeUser,
 		})
 		defer closer.Close()
 
@@ -169,7 +168,7 @@ func TestProvisionerDaemonServe(t *testing.T) {
 			FileID:        file.ID,
 			Provisioner:   codersdk.ProvisionerTypeEcho,
 			ProvisionerTags: map[string]string{
-				provisionerdserver.TagScope: provisionerdserver.ScopeUser,
+				provisionersdk.TagScope: provisionersdk.ScopeUser,
 			},
 		})
 		require.NoError(t, err)
@@ -178,7 +177,7 @@ func TestProvisionerDaemonServe(t *testing.T) {
 		another, _ := coderdtest.CreateAnotherUser(t, client, user.OrganizationID)
 		_ = closer.Close()
 		closer = coderdtest.NewExternalProvisionerDaemon(t, another, user.OrganizationID, map[string]string{
-			provisionerdserver.TagScope: provisionerdserver.ScopeUser,
+			provisionersdk.TagScope: provisionersdk.ScopeUser,
 		})
 		defer closer.Close()
 		workspace := coderdtest.CreateWorkspace(t, another, user.OrganizationID, template.ID)
@@ -204,7 +203,7 @@ func TestProvisionerDaemonServe(t *testing.T) {
 				codersdk.ProvisionerTypeEcho,
 			},
 			Tags: map[string]string{
-				provisionerdserver.TagScope: provisionerdserver.ScopeOrganization,
+				provisionersdk.TagScope: provisionersdk.ScopeOrganization,
 			},
 			PreSharedKey: "provisionersftw",
 		})
@@ -260,7 +259,7 @@ func TestProvisionerDaemonServe(t *testing.T) {
 					codersdk.ProvisionerTypeEcho,
 				},
 				Tags: map[string]string{
-					provisionerdserver.TagScope: provisionerdserver.ScopeOrganization,
+					provisionersdk.TagScope: provisionersdk.ScopeOrganization,
 				},
 				PreSharedKey: "provisionersftw",
 			})
@@ -338,7 +337,7 @@ func TestProvisionerDaemonServe(t *testing.T) {
 				codersdk.ProvisionerTypeEcho,
 			},
 			Tags: map[string]string{
-				provisionerdserver.TagScope: provisionerdserver.ScopeOrganization,
+				provisionersdk.TagScope: provisionersdk.ScopeOrganization,
 			},
 			PreSharedKey: "the wrong key",
 		})
@@ -369,7 +368,7 @@ func TestProvisionerDaemonServe(t *testing.T) {
 				codersdk.ProvisionerTypeEcho,
 			},
 			Tags: map[string]string{
-				provisionerdserver.TagScope: provisionerdserver.ScopeOrganization,
+				provisionersdk.TagScope: provisionersdk.ScopeOrganization,
 			},
 		})
 		require.Error(t, err)
@@ -398,7 +397,7 @@ func TestProvisionerDaemonServe(t *testing.T) {
 				codersdk.ProvisionerTypeEcho,
 			},
 			Tags: map[string]string{
-				provisionerdserver.TagScope: provisionerdserver.ScopeOrganization,
+				provisionersdk.TagScope: provisionersdk.ScopeOrganization,
 			},
 			PreSharedKey: "provisionersftw",
 		})
